+++ src/browser/binds/plugins/npm.ts
@@ -1,5 +1,4 @@
-import { PluginExport } from "../binds/index.js";
-import { Plugins, ScanMethod, Methodology } from "../index.js"
+import { Plugins, ScanMethod, Methodology } from "../../index.js"
 import getValue from "get-value";
 
 export const id = "npm"
@@ -77,4 +76,4 @@ export const methodology: Methodology[] = [
 
 
 const bind: Plugins.TargetBind = { id, name, methodology, testCommand }
-export default ({ viewignored: { addTargets: [bind] } } as PluginExport)
+export default ({ viewignored: { addTargets: [bind] } } as Plugins.PluginExport)
